Move lock validation tests to core lock module.
